added current_grow view and template, returning all current grows for current user
