+++ temp_test/models/dataset.py
@@ -3,8 +3,7 @@ from django.utils import timezone
 from .grow import Grow
 
 class Dataset(models.Model):
-	grow = models.CharField(max_length=100)
-	#grow = models.ForeignKey(Grow, on_delete=models.CASCADE)
+	grow = models.ForeignKey(Grow, on_delete=models.CASCADE)
 	inside_temperature = models.CharField(max_length=50)
 	inside_humidity = models.CharField(max_length=50)
 	outside_temperature = models.CharField(max_length=50)
@@ -15,4 +14,4 @@ class Dataset(models.Model):
 	date = models.DateTimeField(default=timezone.now)
 
 	def __str__(self):
-		return self.grow
+		return f'{self.grow}'
