+++ temp_test/models/grow.py
@@ -11,7 +11,7 @@ class Grow(models.Model):
 	end_date = models.DateTimeField(auto_now=False, auto_now_add=False)
 	description = models.CharField(max_length=200)
 	condition = models.ForeignKey(Condition, on_delete=models.CASCADE)
-	#user = models.ForeignKey(User, on_delete=models.CASCADE)
+	user = models.ForeignKey(User, on_delete=models.CASCADE)
 
 	def __str__(self):
 		return self.name
