+++ temp_test/templates/registration/login.html
@@ -1,12 +1,7 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-	<meta charset="UTF-8">
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta http-equiv="X-UA-Compatible" content="ie=edge">
-	<title>Document</title>
-</head>
-<body>
+{%  extends 'temp_test/index.html' %}
+
+{% block content %}
+<div>
 	{% if form.errors%}
 	<p>There's an error</p>
 	{%  endif %}
@@ -24,6 +19,5 @@
 	<form method='get' action='/register'>
 		<input type="submit" = value="register">
 	</form>
-
-</body>
-</html>
+</div>
+{% endblock %}
