+++ temp_test/templates/temp_test/home.html
@@ -4,6 +4,14 @@
 {% block content %}
 <div>
 	{% include 'temp_test/nav.html' with nav_location=view.location %}
-	<h1>Home</h1>
+	<form>
+		<input type='submit' value='Start New Grow'>
+	</form>
+	<form>
+		<input type='submit' value='View History'>
+	</form>
+	<form method='get' action='/current_grows'>
+		<input type='submit' value='Go To Current Grows'>
+	</form>
 </div>
 {% endblock content %}
