+++ temp_test/urls.py
@@ -3,8 +3,8 @@ from . import views
 
 urlpatterns = [
     path('', views.index, name='index'),
-	path('run/', views.run, name='run'),
-	path('stats/', views.stats, name='stats'),
+	path('start_grow/', views.start_grow, name='start_grow'),
 	path('register/', views.register, name='register'),
-	path('logout/', views.user_logout, name='logout')
+	path('logout/', views.user_logout, name='logout'),
+	path('current_grows/', views.current_grows, name='current_grows')
 ]
