+++ temp_test/views/__init__.py
@@ -1,9 +1,8 @@
-#from .type import TypeView
-#from .category import CategoryView
-#from .variety import VarietyView
-#from .condition import ConditionView
-#from .grow import GrowView
-from .dataset import *
-#from .run import run_program
+from .index import index
+from .register import register
+from .start_grow import start_grow
+from .user_logout import user_logout
+from .current_grows import current_grows
+
