+++ temp_test/views/user_logout.py
@@ -1,0 +1,7 @@
+from django.shortcuts import render, redirect
+from django.contrib.auth import logout
+
+
+def user_logout(request):
+	logout(request)
+	return redirect('login')
